cleaned up files. Added in some icons and functionality to switch icons.
